moved tests andstories to component folder
