+++ src/components/Thing/Thing.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, Story } from '@storybook/react';
-import { Thing, Props } from '../src';
+import { Thing, Props } from './index';
 
 const meta: Meta = {
   title: 'Welcome',
