+++ src/components/Thing/index.tsx
